moved queries
