+++ src/containers/MyNdla/arenaQueries.ts
@@ -7,7 +7,12 @@
  */
 
 import { gql } from '@apollo/client';
-import { GQLArenaUserQuery } from '../../graphqlTypes';
+import {
+  GQLArenaUserQuery,
+  GQLArenaPageQuery,
+  GQLArenaCategoryQuery,
+  GQLArenaTopicByIdQuery,
+} from '../../graphqlTypes';
 import { useGraphQuery } from '../../util/runQueries';
 
 const arenaUserFragment = gql`
@@ -21,6 +26,63 @@ const arenaUserFragment = gql`
   }
 `;
 
+const arenaCategoriesFragment = gql`
+  fragment ArenaCategoriesFragment on ArenaCategory {
+    __typename
+    description
+    disabled
+    htmlDescription
+    id
+    name
+    postCount
+    slug
+  }
+`;
+
+const arenaCategoryFragment = gql`
+  fragment ArenaCategoryFragment on ArenaCategory {
+    __typename
+    description
+    disabled
+    htmlDescription
+    id
+    name
+    postCount
+    slug
+  }
+`;
+
+const arenaTopicFragment = gql`
+  fragment ArenaTopicFragment on ArenaTopic {
+    __typename
+    categoryId
+    id
+    locked
+    postCount
+    slug
+    timestamp
+    title
+  }
+`;
+
+const arenaPostFragment = gql`
+  fragment ArenaPostFragment on ArenaPost {
+    __typename
+    content
+    id
+    timestamp
+    topicId
+    isMainPost
+    user {
+      displayName
+      groupTitleArray
+      profilePicture
+      username
+      location
+    }
+  }
+`;
+
 export const arenaUserQuery = gql`
   query ArenaUser($username: String!) {
     arenaUser(username: $username) {
@@ -30,9 +92,71 @@ export const arenaUserQuery = gql`
   ${arenaUserFragment}
 `;
 
+export const arenaCategoriesQuery = gql`
+  query arenaPage {
+    arenaCategories {
+      ...ArenaCategoriesFragment
+    }
+  }
+  ${arenaCategoriesFragment}
+`;
+
+export const arenaCategoryQuery = gql`
+  query arenaCategory($categoryId: Int!, $page: Int!) {
+    arenaCategory(categoryId: $categoryId, page: $page) {
+      ...ArenaCategoryFragment
+      topics {
+        ...ArenaTopicFragment
+      }
+      topicCount
+    }
+  }
+  ${arenaCategoryFragment}
+  ${arenaTopicFragment}
+`;
+
+export const arenaTopicById = gql`
+  query arenaTopicById($topicId: Int!, $page: Int!) {
+    arenaTopic(topicId: $topicId, page: $page) {
+      ...ArenaTopicFragment
+      posts {
+        ...ArenaPostFragment
+      }
+    }
+  }
+  ${arenaTopicFragment}
+  ${arenaPostFragment}
+`;
+
 export const useArenaUser = (username: string) => {
   const { data } = useGraphQuery<GQLArenaUserQuery>(arenaUserQuery, {
     variables: { username },
   });
   return { arenaUser: data?.arenaUser };
+};
+
+export const useArenaCategories = () => {
+  const { data, loading, error } =
+    useGraphQuery<GQLArenaPageQuery>(arenaCategoriesQuery);
+  return { arenaCategories: data?.arenaCategories, loading, error };
+};
+
+export const useArenaCategory = (categoryId: number, page: number) => {
+  const { data, loading, error } = useGraphQuery<GQLArenaCategoryQuery>(
+    arenaCategoryQuery,
+    {
+      variables: { categoryId, page },
+    },
+  );
+  return { arenaCategory: data?.arenaCategory, loading, error };
+};
+
+export const useArenaTopic = (topicId: number, page: number) => {
+  const { data, loading, error } = useGraphQuery<GQLArenaTopicByIdQuery>(
+    arenaTopicById,
+    {
+      variables: { topicId, page },
+    },
+  );
+  return { arenaTopic: data?.arenaTopic, loading, error };
 };
